% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{AlRich}
\alias{AlRich}
\title{Calculate Allelic Richness}
\usage{
AlRich(datafile = data.frame, n = matrix)
}
\arguments{
\item{datafile}{A data frame containing the data as read in by \link{LoadData}}

\item{n}{A matrix representing the number of markers successfully genotyped like the output of the \link{TypedLoci} function.}
}
\value{
A vector containing the allelic richness for each locus.
}
\description{
This function calculates allelic richness based on the provided genetic data.
}
