% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{BootStrap3}
\alias{BootStrap3}
\title{Perform Bootstrap Analysis}
\usage{
BootStrap3(A = data.frame, Rep = 20, Stat = 1)
}
\arguments{
\item{A}{A data frame containing the data as read in by \link{LoadData}}

\item{Rep}{Number of bootstrap replicates.}

\item{Stat}{Type of statistic to compute (1 for AlRich, 2 for TwoLevelGST)}
}
\value{
Either a matrix of AlRich statistics or a list containing various statistics computed using TwoLevelGST.
}
\description{
This function performs bootstrap analysis on genetic data.
}
