% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_version.R
\name{check_fsa_v_batch}
\alias{check_fsa_v_batch}
\title{Check .fsa Version and Batch Information}
\usage{
check_fsa_v_batch(x)
}
\arguments{
\item{x}{The path to the folder from the current directory where the .fsa
files that will be analyzed are stored.}
}
\value{
A written summary of how many .fsa files are in the folder and which
version they are.
}
\description{
This function analyzes .fsa files in a specified folder, providing a summary
of their version and batch information.
}
\examples{
file_path <- system.file("extdata", package = "pooledpeaks")
check_fsa_v_batch(x = file_path)

}
