% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{DistCor}
\alias{DistCor}
\title{Distance Correlation}
\usage{
DistCor(GD = matrix)
}
\arguments{
\item{GD}{A matrix containing the genetic distance data.}
}
\value{
A plot showing the Expected Genetic Distance versus Realized Genetic Distance
}
\description{
Calculate the correlation between expected and realized genetic distances and plot them.
}
