% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{EmpiricalSE}
\alias{EmpiricalSE}
\title{Calculate Empirical Standard Error}
\usage{
EmpiricalSE(datafile = data.frame, NLoci = 10)
}
\arguments{
\item{datafile}{A data frame containing genetic data from the \link{LoadData}}

\item{NLoci}{Number of loci to sample in each iteration.}
}
\value{
A numeric vector containing the empirical standard error estimates.
}
\description{
This function calculates the empirical standard error based on repeated sampling.
}
