% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{GeneIdentityMatrix}
\alias{GeneIdentityMatrix}
\title{Gene Identity Matrix}
\usage{
GeneIdentityMatrix(RawData = data.frame, LociGenotyped = matrix)
}
\arguments{
\item{RawData}{A data frame containing the input data must be in LoadData style \link{LoadData}.}

\item{LociGenotyped}{The Output from the TypedLoci function}
}
\value{
The Gene Identity Matrix
}
\description{
Using the number of typed loci, this function calculates the gene identity between all possible pairwise combinations between individuals for all markers creating a matrix.
}
