% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{GeneticDistanceMatrix}
\alias{GeneticDistanceMatrix}
\title{Genetic Distance Matrix}
\usage{
GeneticDistanceMatrix(J = matrix)
}
\arguments{
\item{J}{The Gene Identity Matrix created using \link{GeneIdentityMatrix}}
}
\value{
The Genetic Distance Matrix
}
\description{
This function calculates the genetic distance matrix from a given gene identity matrix.
}
\examples{
gene_identity_matrix <- matrix(c(
0.3164550, 0.2836333, 0.2760485,
0.2836333, 0.3106084, 0.2867215,
0.2760485, 0.2867215, 0.3338663
), nrow = 3, byrow = TRUE,
dimnames = list(paste0("Sample", 1:3), paste0("Sample", 1:3)))

GeneticDistanceMatrix(gene_identity_matrix)
}
