% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate_check.R
\name{JostD_KK}
\alias{JostD_KK}
\title{Pairwise Jost D between replicates}
\usage{
JostD_KK(Ni1, Ni2)
}
\arguments{
\item{Ni1}{Vector containing the allele frequencies of the first duplicate sample.}

\item{Ni2}{Vector containing the allele frequencies of the second duplicate sample.}
}
\value{
The calculated Jost's D value.
}
\description{
This function calculates Jost's D between two columns, specifically designed for comparing duplicate samples based on allele frequencies.
}
