% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_to_tdf.R
\name{lf_to_tdf}
\alias{lf_to_tdf}
\title{Transform LF to TDF}
\usage{
lf_to_tdf(x)
}
\arguments{
\item{x}{A data frame in LF format ideally coming out of the clean_scores function.}
}
\value{
A transformed data frame in TDF format.
}
\description{
This function transforms a data frame from LF (long format) to TDF (table format),
performing various data manipulation steps such as spreading data across columns,
removing NA and/or 0 columns, merging ID allele heights within each replicate,
transposing the table, converting from character to numeric class, and replacing
empty data with "0".
}
