% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{LoadData}
\alias{LoadData}
\title{Load Genetic Data}
\usage{
LoadData(datafile = NULL)
}
\arguments{
\item{datafile}{The path to your datafile. The format of your data should be .txt or .csv.}
}
\value{
A data frame containing the imported data formatted in the way necessary for downstream population genetic functions.
}
\description{
This function imports data for genetic analysis.
}
