% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{MDSplot}
\alias{MDSplot}
\title{Multi Dimensional Scaling (MDS) Plot}
\usage{
MDSplot(
  distance = matrix,
  pcs = c(1, 2),
  PF = NULL,
  y = c("dodgerblue", "red", "turquoise3", "purple", "olivedrab3")
)
}
\arguments{
\item{distance}{A matrix containing the genetic distance data.}

\item{pcs}{A numeric vector specifying the principal coordinates to plot.}

\item{PF}{A factor vector specifying population labels.}

\item{y}{A character vector specifying colors for population labels.}
}
\value{
The ouput is the MDS plot for the samples for the specified principal coordinates.
}
\description{
Generate a multidimensional scaling (MDS) plot from genetic distance data.
}
