% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledpeaks-package.R
\docType{package}
\name{pooledpeaks-package}
\alias{pooledpeaks}
\alias{pooledpeaks-package}
\title{pooledpeaks: Genetic Analysis of Pooled Samples}
\description{
Analyzing genetic data obtained from pooled samples. This package can read in Fragment Analysis output files, process the data, and score peaks, as well as facilitate various analyses, including cluster analysis, calculation of genetic distances and diversity indices, as well as bootstrap resampling for statistical inference. Specifically tailored to handle genetic data efficiently, researchers can explore population structure, genetic differentiation, and genetic relatedness among samples. We updated some functions from Covarrubias-Pazaran et al. (2016) \doi{10.1186/s12863-016-0365-6} to allow for the use of new file formats and referenced the following to write our genetic analysis functions: Long et al. (2022) \doi{10.1038/s41598-022-04776-0}, Jost (2008) \doi{10.1111/j.1365-294x.2008.03887.x}, Nei (1973) \doi{10.1073/pnas.70.12.3321}, Foulley et al. (2006) \doi{10.1016/j.livprodsci.2005.10.021}, Chao et al. (2008) \doi{10.1111/j.1541-0420.2008.01010.x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kmkuesters/pooledpeaks}
  \item Report bugs at \url{https://github.com/kmkuesters/pooledpeaks/issues}
}

}
\author{
\strong{Maintainer}: Kathleen Kuesters \email{kathleen.kuesters@gmail.com} (\href{https://orcid.org/0000-0003-0238-7889}{ORCID}) [copyright holder]

}
\keyword{internal}
