% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replicate_check.R
\name{Rep_check}
\alias{Rep_check}
\title{Replicate Check for Duplicate Samples}
\usage{
Rep_check(df)
}
\arguments{
\item{df}{The input data frame containing peak heights for each sample.}
}
\value{
A data frame containing the average peak heights for each sample, with flagged samples where duplicates have a Jost's D exceeding 0.05.
}
\description{
This function checks for duplicate samples in the input data frame and calculates the average peak heights for each sample. If the Jost's D between duplicate samples exceeds 0.05, it flags those samples.
}
