% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{RWCDistanceMatrix}
\alias{RWCDistanceMatrix}
\title{Random Walk Covariance Distance Matrix}
\usage{
RWCDistanceMatrix(J = matrix)
}
\arguments{
\item{J}{The Genetic Distance Matrix calculated using \link{GeneticDistanceMatrix}}
}
\value{
A matrix representing the distance matrix calculated using the Random Walk Covariance method.
}
\description{
This function calculates the RWC (Random Walk Covariance) distance matrix from a given matrix of genetic distances.
}
