% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{SampleOfLoci}
\alias{SampleOfLoci}
\title{Sample Of Loci}
\usage{
SampleOfLoci(aaax = data.frame, NLoci = max(aaax[, 1]))
}
\arguments{
\item{aaax}{A data frame containing the input data must be in LoadData style \link{LoadData}.}

\item{NLoci}{An integer specifying the number of loci to sample.}
}
\value{
A data frame containing the sampled loci.
}
\description{
An internal function that supports ClusterFromSamples. Sample loci from a dataset based on the number of loci specified.
}
