% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_marker_wrap.R
\name{score_markers_rev3}
\alias{score_markers_rev3}
\title{Score Markers Wrapper}
\usage{
score_markers_rev3(
  my.inds,
  channel = 1,
  n.inds = NULL,
  panel = NULL,
  shift = 0.8,
  ladder,
  channel.ladder = NULL,
  ploidy = 2,
  left.cond = c(0.6, 3),
  right.cond = 0.35,
  warn = FALSE,
  windowL = 0.5,
  windowR = 0.5,
  init.thresh = 200,
  ladd.init.thresh = 200,
  method = "iter2",
  env = parent.frame(),
  my.palette = NULL,
  plotting = FALSE,
  plotdir = "plots_scoring",
  pref = 3
)
}
\arguments{
\item{shift}{All peaks at that distance from the tallest peak will be ignored and be considered noise.}

\item{windowL}{the window means that all peaks closer by that distance to the left of the panel peaks will be accounted as peaks.}

\item{windowR}{the window means that all peaks closer by that distance to the right of the panel peaks will be accounted as peaks.}

\item{pref}{}
}
\description{
This is a revision of the Fragman script score.markers, for the original instructions and parameters, run '?score.markers'.
This revision designates separate parameters for Left and Right search windows.
}
