% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{TypedLoci}
\alias{TypedLoci}
\title{Typed Loci}
\usage{
TypedLoci(datafile = data.frame)
}
\arguments{
\item{datafile}{A data frame containing the input data must be in LoadData style \link{LoadData}.}
}
\value{
A matrix representing processed data.
}
\description{
This function calculates the number of loci successfully genotyped by each individual included in our data set
}
