% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster}
\alias{cluster}
\title{K-means Clustering}
\usage{
cluster(RawData = data.frame, K = 2)
}
\arguments{
\item{RawData}{A data frame containing the raw data as read in by \link{LoadData}}

\item{K}{An integer specifying the number of clusters.}
}
\value{
A list containing the results of the K-means cluster analysis, including cluster assignments and original data.
}
\description{
K-means Clustering
}
