% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{ClusterFromSamples}
\alias{ClusterFromSamples}
\title{Cluster From Samples}
\usage{
ClusterFromSamples(datafile = data.frame, numloci = 5, reps = 100)
}
\arguments{
\item{datafile}{A data frame containing the data.}

\item{numloci}{An integer specifying the number of loci to sample.}

\item{reps}{An integer specifying the number of repetitions.}
}
\value{
A matrix containing statistics calculated from the clustering results.
}
\description{
Perform clustering on samples of loci from a data frame and calculate statistics.
}
