% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_setup.R
\name{GeneticDistanceMatrix}
\alias{GeneticDistanceMatrix}
\title{Genetic Distance Matrix}
\usage{
GeneticDistanceMatrix(J = matrix)
}
\arguments{
\item{J}{The Gene Identity Matrix created using \link{GeneIdentityMatrix}}
}
\value{
The Genetic Distance Matrix
}
\description{
This function calculates the genetic distance matrix from a given gene identity matrix.
}
