% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{preGST}
\alias{preGST}
\title{Pre GST Calculation}
\usage{
preGST(G = matrix)
}
\arguments{
\item{G}{The gene identity matrix}
}
\value{
The GST value.
}
\description{
This function calculates the GST from a gene identity matrix.
}
