% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Gen_Functions.R
\name{preJostD}
\alias{preJostD}
\title{Calculate Pre-Jost's D}
\usage{
preJostD(G = matrix)
}
\arguments{
\item{G}{A square matrix representing a genetic distance matrix.}
}
\value{
The Jost's D value.
}
\description{
This function calculates the pre-Jost's D measure from a genetic distance matrix.
}
